Patch #4995, bugfix #4985
Patch for this bug provided by Johannes Meng.
Here's his log:

When using the mouse wheel on an inverted 3d view (that is back, left or
bottom) the name display changed to front (right, top). Also when using
mouse wheel with modifier keys (strg/shift) on a normal view, it changed
the display to the inverted variant.
What I did was check the type of event that is handeled before
setting/unsetting V3D_OPP_DIRECTION_NAME. I simply added two if statements
and expanded the comment by a line explaining it a bit.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@8504 954f8c5b-7b00-dc11-b283-0030488c597c