Bugfix #5146
Copy modifiers menu should be available for Lattice and Surface objects.

Note it does not support copying between different object types, that
is being indicated with a notice popup now.

I rather not code copying to other object types for now, this is something
which fits better in the modifiers API first. Right now the buttons code
handles this.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@8743 954f8c5b-7b00-dc11-b283-0030488c597c