->Join triangles fixes.
There was a bug with the new join triangles code that caused concave faces to be
created. This is fixed now and is tweakable using the the 'threshold' value set
in the 'Mesh Tools' panel in editbuttons.

Also removed the popup notice telling you how many triangles had been joined. It
was only there for debugging purposes.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@8785 954f8c5b-7b00-dc11-b283-0030488c597c