Lightgroups!
Functionality was a bit limited still; now added the option to bind
a group of Lamps entirely to a Material, excluding them from any other
Material. Note that Lamp visibility layer options still work as for
Lamps usually.

In preview.blend: made a lightgroup for all lamps, and set them to be
the "RenderLayer override". That way it will ignore local material
lightgroups for previews.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@9249 954f8c5b-7b00-dc11-b283-0030488c597c