== VisualRot Keyframing ==
After extensive testing, I've concluded that VisualRot keyframing
didn't work at all for bones. On the other hand, VisualRot for Objects
worked fine. Now I've come up with a fix, which should make this
actually work for IK/FK switching (I hope).

The fix I'm committing is slightly hacky as it tempolarily overwrites the
bone's rotation/quat values with the 'visual' ones, inserts keyframes,
then resets the values it wrote on before.

This hack was needed as for some strange reason, the values being
added as keyframes were not the ones that had been derived. Note
that on some bones, there will be a little bit of strange rotation after
frame changes while their constraints are left on.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@9767 954f8c5b-7b00-dc11-b283-0030488c597c