Bugfix #5316
Order of freeing of Blender database is important, but the past year(s)
new database entries were added in wrong places, so freeing could go in
wrong order. In this case it was Groups being freed before Objects. But
there are more errors...

It only shows really in Windows, since it's the OS that is sensitive for
writing data in freed memory. For other OSes it doesn't make Blender more
stable, only mem-frees happened, no mem-allocs.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@9823 954f8c5b-7b00-dc11-b283-0030488c597c