Bugfix for #5896 - Array constant offset not behaving as expected This is not a bug, but rather a misunderstanding of the relationship between object scaling and the Array modifier's constant offset.
The Array modifier's constant offset is specified in local coordinates. In the supplied blend file, the object is scaled up in the x direction by 2.45, and the offset is a constant 2.5 in the x direction. Thus, the final world space offset is 2.45 * 2.5 = 6.125 units in the object's local x direction, as can be seen in the blend file.

I have updated the tooltips for the Array modifier to indicate that offsets are given in local coordinates. I will investigate providing a "Use World Coordinates" toggle button after release.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@9990 954f8c5b-7b00-dc11-b283-0030488c597c