Scripts: - Fixing bug #5950 reported by Stewart Andreason (thanks): http://projects.blender.org/tracker/?func=detail&atid=125&aid=5950&group_id=9
Importing .ac models created Blender objects with user count equal to 2, instead of 1 (so they couldn't be deleted in Blender), because the script kept a second reference to each created Blender object and so Python didn't deallocate things properly at the end of the script.

Now these extra references are deleted in the script, preventing the problem. But this is surely something to fix in the API itself.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@10063 954f8c5b-7b00-dc11-b283-0030488c597c