the input fields for data name had an inconsistant limit for input fields, making it hard to fix problems with library linking when a name changed. some were 18, most 19, and others 21. made all 21 since this is the real limit.
Also new image name limit length of input field to 21 (was 255 but shortened to 21)

The one place this could be useful is if somebody names a metaball with a 21 char name, the copy will not use the motherball.
but this is not as bad as having to use the python console for fixing library linking problems.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@10499 954f8c5b-7b00-dc11-b283-0030488c597c