* Create autosave dir on windows if it doesn't exist
For a long time, Blender has shipped with a .B.blend with a default .B.blend
that has "/tmp" as a default temp directory. This, of course, doesn't usually 
exist on Windows, so anyone who used Blender on Windows and didn't know 
to manually update it, lost all auto save files.

Now on Windows, Blender checks to see if the directory in U.tmpdir exists, 
and if it doesn't, it will create a new 'autosave' dir underneath the .blender
home directory and save auto save files in there.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@10535 954f8c5b-7b00-dc11-b283-0030488c597c