== Bug Fixes ==
[ #6620 ] Problems to select faceloops in Select Only Visible mode
[ #6131 ] Vertex snapping broken in shaded face + edge select mode

Second bug was already "fixed", but that caused the other problem. Forced me to look more deeply into it find the real cause. So, faceloop problem is fixed by revert, the other is fixed by drawing vertice selection colors when in transform (previously was only when vert select was on).


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@10632 954f8c5b-7b00-dc11-b283-0030488c597c