Curves in Editmode:
HKEY and Shift HKEY are used to change the handle type of curve control points in the IPO and Action Editors, and also for curves in the 3d view until recently.

This commit reverses the hotkey changes in:
http://projects.blender.org/pipermail/bf-blender-cvs/2007-May/009879.html

For curves, I've changed the show/hide points hotkeys to:
* Ctrl Alt HKEY  == Hide Point(s)
* Alt HKEY == Show All Points

These aren't optimal, but at least there are usable hotkeys for this for now. Before, there had been a long-standing hotkey conflict. Better hotkeys can be designed as part of 2.5 work.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@10652 954f8c5b-7b00-dc11-b283-0030488c597c