== Action Editor - Bugfix #6660 ==
When trying to expand the IPO-curves for Shapekey Action Channels, Blender crashed due to that being a previously un-handled case.

For some reason, the names of IPO curve channels are still not drawing on some machines.

I've also disabled the display of shapekey channels when there is an action as the two used to overlap and muck up the drawing matrices of each other.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@10677 954f8c5b-7b00-dc11-b283-0030488c597c