After chatting with Broken about grs comments on  the default behavior of the nodes, I realized I had strayed from the path of enlightened blending by causing the math node to create an output the size of the larger of the two inputs.
It has been corrected create the output the size of the 
first image, and in its abscense the second image.  In the event of 
nether input containing image data the node does not function.  I also 
added some early out checks at the beginning of the function to speed it 
up a tad in these cases and commented the code a bit more.  




git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@11224 954f8c5b-7b00-dc11-b283-0030488c597c