== IPO Cleaning - Refactor ==
I've finally come to the conclusion that the old code used to clean up IPO-curves was utterly unsafe/unstable. It used to have a lot of memory-related bugs early on, which I had resolved, but still I found a file where this broke quite badly (thanks Cessen).

Now the code should be saner, as it does away with that evil memory copying that seems to be at fault. It would be advisable to have this properly tested for a few days before considering it for inclusion in 2.45 (stable) branch. 

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@11511 954f8c5b-7b00-dc11-b283-0030488c597c