Bugfix #6923:
NLA was calculating extensions of strips in the wrong order. This was not an issue in many cases, although it is highlighted in the following case:

[strip 1 - 'hold' is on] ----------------------------
                               [strip 2        ]

Logically, the end of strip 1 would hold on until the start of strip 2. However, the start of strip 2 was extended to the end of strip 1 instead.

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@11982 954f8c5b-7b00-dc11-b283-0030488c597c