Bugfix #7294:
When resizing an Outliner window, the contents would eventually get 'pushed out of view' when shrinking the view. 

I've added a bit of a 'hack', which will ensure that this doesn't happen, by forcing the view to look at the left-side of the outliner tree, when the width of the outliner window decreses due to resizing. 

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@11989 954f8c5b-7b00-dc11-b283-0030488c597c